fix ternary expression throwing NPE in some java versions

+++ additive_animations/src/main/java/at/wirecube/additiveanimations/additive_animator/BaseAdditiveAnimator.java
@@ -199,7 +199,9 @@ public abstract class BaseAdditiveAnimator<T extends BaseAdditiveAnimator, V ext
      * This method is for internal use only (keeping track of chained `animateBy` calls).
      */
     protected Float getQueuedPropertyValue(String propertyName) {
-        return mRunningAnimationsManager == null ? 0 : mRunningAnimationsManager.getQueuedPropertyValue(propertyName);
+        // important: you have to return a boxed Float in both branches of the ternary expression, otherwise the Android java compiler
+        // will infer that we want to return a primitive, and auto-unbox the Float - which might result in an NPE!
+        return mRunningAnimationsManager == null ? Float.valueOf(0) : mRunningAnimationsManager.getQueuedPropertyValue(propertyName);
     }
 
     void applyChanges(List<AccumulatedAnimationValue<V>> accumulatedAnimations) {
